server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gabriel-server"
version = "4.1"
description = "Server for the Gabriel real-time AI orchestration framework"
requires-python = ">=3.10"

dependencies = [
    "gabriel-protocol==4.1",
    "websockets",
    "pyzmq",
    "prometheus_client",
    "protobuf>=6.0.0,<7.0.0",
]

authors = [
    {name = "Aditya Chanana", email = "achanana@cs.cmu.edu"},
]

maintainers = [
    {name = "CMU Satyalab", email = "gabriel@cmusatyalab.org"},
]

license = "Apache-2.0"

classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "http://gabriel.cs.cmu.edu"
Repository = "https://github.com/cmusatyalab/gabriel/"
